refactor(settings/theme): cleanup ThemeUtils
