+++ app/src/main/java/org/schulcloud/mobile/utils/SingletonHolder.kt
@@ -1,0 +1,25 @@
+package org.schulcloud.mobile.utils
+
+open class SingletonHolder<out T, in A>(creator: (A) -> T) {
+    private var creator: ((A) -> T)? = creator
+    @Volatile
+    private var instance: T? = null
+
+    fun getInstance(arg: A): T {
+        val i = instance
+        if (i != null)
+            return i
+
+        return synchronized(this) {
+            val i2 = instance
+            if (i2 != null)
+                i2
+            else {
+                val created = creator!!(arg)
+                instance = created
+                creator = null
+                created
+            }
+        }
+    }
+}
